progress towards adding experiments.
Also in process of rewriting code so configurations are
handled last, in actual experiments rather than in
dataset construction. Also adding learning rate search,
data augmentations and looking into lr schedulers.
Depending on the granularity of lr search, and including
experiment repetition, the experiments for object detection
may take long due to the inference time of a model being in
the order of 1 second. We'll see though.